+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 - `RadioService.incrementDonatedStoyCoins`: Method to atomically increment StoyCoins donations for a radio station
 
+**feat(StoyCoins):** serializable model for stoycoins generic
+
 ## 21.7.4
 
 **feat(StoyCoins):** serializable model for stoycoins transactionsv2